% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\name{cellTypeTrain}
\alias{cellTypeTrain}
\alias{cellTypeTrain,scClassifyTrainModel-method}
\title{Accessors of cellTypeTrain for scClassifyTrainModel}
\usage{
cellTypeTrain(x)
}
\arguments{
\item{x}{A `scClassifyTrainModel` object.}
}
\value{
cellTypeTrain of the scClassifyTrainModel slot
}
\description{
Methods to access various components of the `scClassifyTrainModel` object.
}
\examples{

data(trainClassExample_xin)
cellTypeTrain(trainClassExample_xin)

}
