% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\name{cellTypeTree}
\alias{cellTypeTree}
\alias{cellTypeTree,scClassifyTrainModel-method}
\title{Accessors of cellTypeTree for scClassifyTrainModel}
\usage{
cellTypeTree(x)
}
\arguments{
\item{x}{A `scClassifyTrainModel` object.}
}
\value{
cellTypeTree of the scClassifyTrainModel slot
}
\description{
Methods to access various components of the `scClassifyTrainModel` object.
}
\examples{

data(trainClassExample_xin)
cellTypeTree(trainClassExample_xin)

}
