% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\name{.scClassifyTrainModel}
\alias{.scClassifyTrainModel}
\title{The scClassifyTrainModel class}
\usage{
.scClassifyTrainModel(
  name,
  cellTypeTree,
  cellTypeTrain,
  features,
  model,
  modelweights,
  metaData
)
}
\arguments{
\item{name}{Name of the training dataset}

\item{cellTypeTree}{A list indicate a cell type tree}

\item{cellTypeTrain}{A vector of cell type in training dataset}

\item{features}{A vector of character indicates the
features that are trained for this data}

\item{model}{A list stored the training model,
including the features that are selected
and the cell expression matrix that are used for training}

\item{modelweights}{A vector of numeric indicates the weights of each model}

\item{metaData}{A DataFrame stored meta data of training model}
}
\value{
A scClassifyTrainModel object
}
\description{
The scClassifyTrainModel class is designed to stored
training model for scClassify
}
\author{
Yingxin Lin
}
