% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\name{features}
\alias{features}
\alias{features,scClassifyTrainModel-method}
\title{Accessors of features for scClassifyTrainModel}
\usage{
features(x)
}
\arguments{
\item{x}{A `scClassifyTrainModel` object.}
}
\value{
features of the scClassifyTrainModel slot
}
\description{
Methods to access various components of the `scClassifyTrainModel` object.
}
\examples{

data(trainClassExample_xin)
features(trainClassExample_xin)

}
