% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningCurve.R
\name{getN}
\alias{getN}
\title{Function to get the required N given by the accuracy and the learning curve model}
\usage{
getN(res, acc = 0.9)
}
\arguments{
\item{res}{model results returned by \code{learning_curve} function}

\item{acc}{accuracy that are quired}
}
\value{
sample size that are required
}
\description{
Function to get the required N given by the accuracy and the learning curve model
}
