% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\name{model}
\alias{model}
\alias{model,scClassifyTrainModel-method}
\title{Accessors of model for scClassifyTrainModel}
\usage{
model(x)
}
\arguments{
\item{x}{A `scClassifyTrainModel` object.}
}
\value{
model of the scClassifyTrainModel slot
}
\description{
Methods to access various components of the `scClassifyTrainModel` object.
}
\examples{

data(trainClassExample_xin)
model(trainClassExample_xin)

}
