% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\name{modelweights}
\alias{modelweights}
\alias{modelweights,scClassifyTrainModel-method}
\title{Accessors of modelweights for scClassifyTrainModel}
\usage{
modelweights(x)
}
\arguments{
\item{x}{A `scClassifyTrainModel` object.}
}
\value{
modelweights of the scClassifyTrainModel slot
}
\description{
Methods to access various components of the `scClassifyTrainModel` object.
}
\examples{

data(trainClassExample_xin)
modelweights(trainClassExample_xin)

}
