% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\name{name}
\alias{name}
\alias{name,scClassifyTrainModel-method}
\title{Accessors of name for scClassifyTrainModel}
\usage{
name(x)
}
\arguments{
\item{x}{A `scClassifyTrainModel` object.}
}
\value{
name of the scClassifyTrainModel slot
}
\description{
Methods to access various components of the `scClassifyTrainModel` object.
}
\examples{

data(trainClassExample_xin)
name(trainClassExample_xin)

}
