% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHOPACH.R
\name{plotCellTypeTree}
\alias{plotCellTypeTree}
\title{To plot cell type tree}
\usage{
plotCellTypeTree(cutree_list, group_level = NULL)
}
\arguments{
\item{cutree_list}{A list indicates the hierarchical cell type tree}

\item{group_level}{Indicate whether plot or not}
}
\value{
A ggplot object visualising the HOPACH tree
}
\description{
To plot cell type tree
}
