% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTestClass.R
\docType{class}
\name{scClassifyTestRes-class}
\alias{scClassifyTestRes-class}
\title{An S4 class to stored training model for scClassify}
\description{
An S4 class to stored training model for scClassify
}
\section{Slots}{

\describe{
\item{\code{testingDataName}}{Name of the testing dataset}

\item{\code{trainingModelName}}{Name of training dataset}

\item{\code{ensembleLabel}}{A vector of character indicates the ensemble predicted label}

\item{\code{ensembleScore}}{A vector of numeric indicates scores of the ensemble predicted label}

\item{\code{ensembleModelWeights}}{A vector of numeric indicates the weights of each model}

\item{\code{predictedLabel}}{A data.frame indicates the predicted labels for each base model
(each column indicates one base model)}

\item{\code{predictedMatrix}}{A list indicates the labels of all level predictions for
each base model (each list indicates one base model)}

\item{\code{resCategories}}{A data.frame indicates the predicted results for each base model
(if true test label is provided)}

\item{\code{featureCombination}}{A vector of character indicates the features and similarity combinations
that are trained for this data}
}}

