% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\docType{class}
\name{scClassifyTrainModel-class}
\alias{scClassifyTrainModel-class}
\title{An S4 class to stored training model for scClassify}
\description{
An S4 class to stored training model for scClassify
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the training dataset}

\item{\code{cellTypeTrain}}{A vector of cell type in training dataset}

\item{\code{cellTypeTree}}{A list indicate a cell type tree}

\item{\code{features}}{A vector of character indicates the features that are trained for this data}

\item{\code{model}}{A list stored the training model, including the features that are selected
and the cell expression matrix that are used for training}

\item{\code{modelweights}}{A vector of numeric indicates the weights of each model}

\item{\code{metaData}}{A DataFrame stored meta data of training model}
}}

