% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\docType{class}
\name{scClassifyTrainModelList-class}
\alias{scClassifyTrainModelList-class}
\title{An S4 class to stored a list of training models from scClassify}
\description{
An S4 class to stored a list of training models from scClassify
}
