% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scClassifyTrainClass.R
\name{scClassifyTrainModelList}
\alias{scClassifyTrainModelList}
\title{The scClassifyTrainModelList class}
\usage{
scClassifyTrainModelList(...)
}
\arguments{
\item{...}{scClassifyTrainModel objects}
}
\value{
A scClassifyTrainModelList object
}
\description{
The scClassifyTrainModelList class
}
\examples{

data("trainClassExample_xin")
data("trainClassExample_wang")
trainClassExampleList <- scClassifyTrainModelList(trainClassExample_xin,
trainClassExample_wang
)

}
