% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainClassExample_wang}
\alias{trainClassExample_wang}
\title{Subset of pretrained model of Wang et al.}
\format{
An object of class \code{scClassifyTrainModel} of length 1.
}
\source{
Wang YJ, Schug J, Won K-J, Liu C, Naji A, Avrahami D, Golson ML &
Kaestner KH (2016) Single cell transcriptomics of the human endocrine
pancreas. Diabetes: db160405
}
\usage{
data(trainClassExample_wang, package = 'scClassify')
}
\description{
An obejct of scClassifyTrainModel for Wang et al.
}
\keyword{datasets}
