% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainClassExample_xin}
\alias{trainClassExample_xin}
\title{Subset of pretrained model of Xin et al.}
\format{
An object of class \code{scClassifyTrainModel} of length 1.
}
\source{
Xin Y, Kim J, Okamoto H, Ni M, Wei Y, Adler C, Murphy AJ, Yancopoulos GD,
Lin C & Gromada J (2016) RNA sequencing of single human islet cells reveals
type 2 diabetes genes. Cell Metab. 24: 608–615
}
\usage{
data(trainClassExample_xin, package = 'scClassify')
}
\description{
An obejct of scClassifyTrainModel for Xin et al.
}
\keyword{datasets}
