% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scClassify_example}
\alias{scClassify_example}
\title{Example data used in scClassify package}
\format{An object of class \code{list} of length 4.}
\usage{
data(scClassify_example, package = 'scClassify')
}
\description{
A list includes expression matrix and cell type of subsets of
wang et al., xin et al.
}
\keyword{datasets}
