% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainClassExample_wang}
\alias{trainClassExample_wang}
\title{Subset of pretrained model of Wang et al.}
\format{An object of class \code{scClassifyTrainModel} of length 1.}
\usage{
data(trainClassExample_wang, package = 'scClassify')
}
\description{
An obejct of scClassifyTrainModel for Wang et al.
}
\keyword{datasets}
